requirements.txt
Flask>=1.1.1
Jinja2>=2.10.3
SQLAlchemy>=1.3.11
gunicorn>=20.0.0
flask_sqlalchemy>=2.0.0
joblib>=1.0.1
scikit-learn>=0.24.1
tensorflow-cpu>=2.4.1